[FLINK-12839][dist] package flink-shaded-netty-tcnative-dynamic into opt/
Please note that there is also a static version of netty-tcnative but we
currently do not distribute it due to licensing issues. Once openSSL completes
its switch to Apache License v2, we can provide this as well and maybe even
make that one default (by putting it into lib/). Since there are to many things
which may go wrong with the dynamically-linked library (based on the system you
run on), we provide this only in opt/.