[FLINK-12647][network] Make partition release on consumption optional
ResultPartitions being released on consumption is no longer a hard-coded behavior. This is a prerequisite for making partitions consumable multiple times, without knowing ahead-of-time how often that might be. To retain the existing behavior a configurable flag was introduced to force the consumption of all release partitions, but this will likely be removed in the future.