[FLINK-12963] [state-processing-api] Introduce OperatorStateSpec wrapper class
This class is a simple container class to represent an operator state
that is either still defined by a BootstrapTransformation, i.e. new
state that has not been written out yet, or an existing OperatorState.

Introducing this class improves readability of the code, instead of
using Eithers and Tuples that would not have clear semantics for the
user.