[FLINK-12963] [state-processing-api] Reduce dependency on SavepointMetadata in runtime components
Previously, we were passing around SavepointMetadata across different
runtime classes, such as MergeOperatorStates operator and
BoostrapTransformation.write().

This isn't necessary, as these runtime classes actually do not need
functionality that the SavepointMetadata provides. The SavepointMetadata
should essentially just be a data structure that maintains what existing
operator states there are, and which operator states are added, in order
to build the job graph.

This commit changes this by only passing in the necessary information
wrapped by SavepointMetadata to the runtime classes. By doing so, we can
avoid making SavepointMetadata a Serializable, since it is now no longer
part of the built job graph.